MAJ depuis la production

+++ wp-content/plugins/google-analyticator/google-analyticator.php
@@ -1,15 +1,18 @@
 <?php
 /*
  * Plugin Name: Google Analyticator
- * Version: 6.4.8
+ * Version: 6.4.9.4
  * Plugin URI: http://www.videousermanuals.com/google-analyticator/?utm_campaign=analyticator&utm_medium=plugin&utm_source=readme-txt
  * Description: Adds the necessary JavaScript code to enable <a href="http://www.google.com/analytics/">Google's Analytics</a>. After enabling this plugin you need to authenticate with Google, then select your domain and you're set.
- * Author: Video User Manuals Pty Ltd
- * Author URI: http://www.videousermanuals.com/?utm_campaign=analyticator&utm_medium=plugin&utm_source=readme-txt
+ * Author: SumoMe
+ * Author URI: http://www.sumome.com/?src=wp_readme
  * Text Domain: google-analyticator
  */
 
-define('GOOGLE_ANALYTICATOR_VERSION', '6.4.8');
+//error_reporting(E_ALL);
+//ini_set('display_errors', '1');
+
+define('GOOGLE_ANALYTICATOR_VERSION', '6.4.9');
 
 define('GOOGLE_ANALYTICATOR_CLIENTID', '1007949979410.apps.googleusercontent.com');
 define('GOOGLE_ANALYTICATOR_CLIENTSECRET', 'q06U41XDXtzaXD14E-KO1hti'); //don't worry - this don't need to be secret in our case
@@ -148,8 +151,8 @@ function add_ga_option_page() {
 
 
 	if(ga_get_active_addons() == false){
-		$plugin_page = add_options_page(__('Google Analyticator Settings', 'google-analyticator'), 'Google Analytics', 'manage_options', basename(__FILE__), 'ga_settings_page');
-		add_action('load-'.$plugin_page, 'ga_pre_load' );
+		//$plugin_page = add_options_page(__('Google Analyticator Settings', 'google-analyticator'), 'Google Analytics', 'manage_options', basename(__FILE__), 'ga_settings_page');
+		//add_action('load-'.$plugin_page, 'ga_pre_load' );
 	}
 	   $activate_page = add_submenu_page( null, 'Activation', 'Google Analytics', 'manage_options', 'ga_activate' , 'ga_activate');
 	   $reset_page = add_submenu_page(null, 'Reset', 'Reset', 'activate_plugins', 'ga_reset', 'ga_reset' );
@@ -260,15 +263,17 @@ function ga_filter_plugin_actions($links) {
 	$new_links = array();
 
 	$new_links[] = '<a href="' . ga_analyticator_setting_url() .'">' . __('Settings', 'google-analyticator') . '</a>';
-        $new_links[] = '<a href="' . admin_url('options-general.php?page=ga_reset">') . __('Reset', 'google-analyticator') . '</a>';
+        $new_links[] = '<a href="' . wp_nonce_url( admin_url('options-general.php?page=ga_reset'), 'ga-reset' ) .'">'. __('Reset', 'google-analyticator') . '</a>';
 
 	return array_merge($new_links, $links);
 }
 
 function ga_do_reset()
 {
-    global $wpdb;
-    
+	global $wpdb;
+	// Check to make sure referer is same as host.
+	check_admin_referer( 'ga-reset' );
+
     // Delete all GA options.
     delete_option(key_ga_status);
     delete_option(key_ga_disable_gasites);
@@ -294,10 +299,10 @@ function ga_do_reset()
     delete_option('ga_google_authtoken');
     delete_option('ga_profileid');
     delete_transient('ga_admin_stats_widget');
-    
+
     // Need to remove cached items from GA widgets 
     $wpdb->query( "delete from $wpdb->options where `option_name` like 'google_stats_visitsGraph_%'");
- 
+
     wp_redirect( admin_url( 'options-general.php?page=ga_activate' ) );
     exit;
 }
@@ -308,6 +313,11 @@ function ga_settings_page(){
 	ga_options_page();
 }
 function ga_options_page() {
+	add_thickbox();
+
+	if (array_key_exists('ga_analyticator_global_notification', $_GET) && $_GET['ga_analyticator_global_notification'] == 0) {
+        update_option('ga_analyticator_global_notification', 0);
+	}
 
 	// If we are a postback, store the options
 	if (isset($_POST['info_update'])) {
@@ -465,6 +475,14 @@ function ga_options_page() {
 		
 	?>
 <div class="wrap">
+    <style type="text/css">
+            #ga_analyticator_global_notification a.button:active {vertical-align:baseline;}
+    </style>
+    <div id="ga_analyticator_global_notification" class="updated" style="border:3px solid #317A96;position:relative;background:##3c9cc2;background-color:#3c9cc2;color:#ffffff;height:70px;">
+        <p style="font-size:16px;line-height:50px;">
+                <?php _e('Grow your site faster!'); ?> &nbsp;<a style="background-color: #6267BE;border-color: #3C3F76;" href="<?php echo admin_url('plugin-install.php?tab=plugin-information&plugin=sumome&TB_iframe=true&width=743&height=500'); ?>" class="thickbox button button-primary">Get SumoMe WordPress Plugin</a>
+        </p>
+    </div>
   <form method="post" action="<?php echo ga_analyticator_setting_url();?>">
     <?php
 			# Add a nonce
@@ -484,7 +502,15 @@ function ga_options_page() {
     <?php
 $addons = get_option("gapro_addons");
 if(!$addons){?>
-    <div id="vumga-sidebar" style="position: absolute; top: 0; right: 0; width: 250px; border: 1px solid #ccc; padding: 20px; background:#FFFFFF"> <a href="http://get.videousermanuals.com/ga-pro/?utm_campaign=analyticatorpro&utm_medium=plugin&utm_source=settings" target="_blank"><img src="<?php echo plugins_url('gapro-plugin-advert-sidebar.png', __FILE__ ); ?>" alt="Learn More" title="Learn More" /></a> </div>
+    <div id="vumga-sidebar" style="position: absolute; top: 40px; right: 0; width: 250px; border: 1px solid #ccc; padding: 20px; background:#FFFFFF"> 
+		<h3 style="text-align:center">Support us</h3>
+		<p>1- Check out our <a target="_blank" href="https://wordpress.org/plugins/sumome">SumoMe plugin</a></p>
+		<p>2- <a target="_blank" href="https://wordpress.org/support/view/plugin-reviews/google-analyticator">Leave a :) Review</a></p>	
+		<p>3- Have a great day!</p>
+		<h3 style="text-align:center">Show off your analytics</h3>
+		<p>Use short code <b>[analytics]</b> anywhere on your site to show your analytics publicly.</p>
+		<p>Use short code <b>[analytics-counter]</b> anywhere on your site to display the page view counter widget.</p>
+	</div>
     <?php }?>
     <div style="margin-right: 320px;">
     <table class="form-table" cellspacing="2" cellpadding="5" width="100%">
@@ -587,6 +613,36 @@ if(!$addons){?>
 					  </p>
 						</td>
       </tr>
+      <tr>
+        <th valign="top" style="padding-top: 10px;"> <label for="<?php echo key_ga_widgets; ?>">
+            <?php _e('Support us', 'google-analyticator'); ?>:</label>
+        </th>
+        <td><?php
+						echo "<select name='".key_ga_show_ad."' id='".key_ga_show_ad."'>\n";
+
+						echo "<option value='1'";
+						if(get_option(key_ga_show_ad) == '1')
+							echo " selected='selected'";
+						echo ">" . __('Yes', 'google-analyticator') . "</option>\n";
+
+						echo "<option value='0' ";
+						if(get_option(key_ga_show_ad) == '0')
+							echo" selected='selected'";
+						echo ">" . __('No', 'google-analyticator') . "</option>\n";
+
+						echo "</select>\n";
+						?>
+          <p  class="setting-description">
+            <?php _e('Show our link on the admin dashboard. Pretty please.', 'google-analyticator'); ?>
+          </p></td>
+      </tr>
+	  <tr>
+			<td><b>View your Dashboard</b></td>
+			<td><a href="../wp-admin/index.php">Click here</a></td>
+	  </tr>
+	<tr>
+		<td><input type="submit" class="button button-primary" name="info_update" value="<?php _e('Save Changes', 'google-analyticator'); ?>" /></td>
+    </tr>
       <tr>
         <td colspan="2" style="padding-left:0"><h3>
             <?php _e('Tracking Settings', 'google-analyticator'); ?>
@@ -933,29 +989,6 @@ if(!$addons){?>
             <?php _e('Disabling this option will completely remove the Dashboard Summary widget and the theme Stats widget. Use this option if you would prefer to not see the widgets.', 'google-analyticator'); ?>
           </p></td>
       </tr>
-      <tr>
-        <th valign="top" style="padding-top: 10px;"> <label for="<?php echo key_ga_widgets; ?>">
-            <?php _e('Display Ad', 'google-analyticator'); ?>:</label>
-        </th>
-        <td><?php
-						echo "<select name='".key_ga_show_ad."' id='".key_ga_show_ad."'>\n";
-
-						echo "<option value='1'";
-						if(get_option(key_ga_show_ad) == '1')
-							echo " selected='selected'";
-						echo ">" . __('Yes', 'google-analyticator') . "</option>\n";
-
-						echo "<option value='0' ";
-						if(get_option(key_ga_show_ad) == '0')
-							echo" selected='selected'";
-						echo ">" . __('No', 'google-analyticator') . "</option>\n";
-
-						echo "</select>\n";
-						?>
-          <p  class="setting-description">
-            <?php _e('You can disable the ad on the admin dashboard.', 'google-analyticator'); ?>
-          </p></td>
-      </tr>
       <tr<?php if(!$useAuth){echo ' style="display:none"';}?>>
         <th valign="top" style="padding-top: 10px;"> <label for="<?php echo key_ga_dashboard_role ?>">
             <?php _e('User roles that can see the dashboard widget', 'google-analyticator'); ?>:</label>
@@ -1265,7 +1298,7 @@ function ga_current_user_is($roles)
 }
 
 function ga_analyticator_setting_url(){
-	return ( ga_get_active_addons() == false) ? admin_url("options-general.php?page=google-analyticator.php") : admin_url("admin.php?page=ga-pro-experiment.php");
+	return ( ga_get_active_addons() == false) ? admin_url("admin.php?page=google-analyticator") : admin_url("admin.php?page=ga-pro-experiment.php");
 }
 
 function ga_get_active_addons(){
@@ -1303,3 +1336,78 @@ function add_ga_admin_footer(){
 			})';
 		echo '</script>';
 }
+
+/**
+ * Shortcode
+ */
+
+//[analytics]
+function ga_analyticator_shortcode( $atts ) {
+	# Include the Google Analytics Summary widget
+	require_once('google-analytics-summary-widget.php');
+
+	ob_start();
+
+	$google_analytics_summary = new GoogleAnalyticsSummary(TRUE);
+	// Wrap it with these divs to mimic the admin dashboard widget structure.
+	echo '<div id="google-analytics-summary"><div class="inside">';
+	$google_analytics_summary->widget();
+	echo '</div></div>';
+
+	return ob_get_clean();
+}
+
+//[analytics-counter]
+function ga_analyticator_counter_shortcode( $atts ) {
+	# Include Google Analytics Stats widget
+	require_once('google-analytics-stats-widget.php');
+	ob_start();
+
+	$google_analytics_widget = new GoogleStatsWidget(TRUE);
+
+	return ob_get_clean();
+}
+
+add_shortcode( 'analytics', 'ga_analyticator_shortcode' );
+add_shortcode( 'analytics-counter', 'ga_analyticator_counter_shortcode' );
+
+add_option('ga_analyticator_global_notification', 1);
+
+function ga_analyticator_global_notice() {
+	if (in_array(substr(basename($_SERVER['REQUEST_URI']), 0, 11), array('plugins.php', 'index.php')) && get_option('ga_analyticator_global_notification') == 1) {
+		?>
+			<style type="text/css">
+				#ga_analyticator_global_notification a.button:active {vertical-align:baseline;}
+			</style>
+			<div class="updated" id="ga_analyticator_global_notification" style="border:3px solid #317A96;position:relative;background:##3c9cc2;background-color:#3c9cc2;color:#ffffff;height:70px;">
+				<a class="notice-dismiss" href="<?php echo admin_url('admin.php?page=google-analyticator&ga_analyticator_global_notification=0'); ?>" style="right:165px;top:0;"></a>
+				<a href="<?php echo admin_url('admin.php?page=google-analyticator&ga_analyticator_global_notification=0'); ?>" style="position:absolute;top:9px;right:15px;color:#ffffff;">Dismiss and go to settings</a>
+				<p style="font-size:16px;line-height:50px;">
+					<?php _e('Grow your site faster!'); ?> &nbsp;<a style="background-color: #6267BE;border-color: #3C3F76;" href="<?php echo admin_url('plugin-install.php?tab=plugin-information&plugin=sumome&TB_iframe=true&width=743&height=500'); ?>" class="thickbox button button-primary">Get SumoMe WordPress Plugin</a>
+				</p>
+	        </div>
+		<?php
+	}
+}
+add_action( 'admin_notices', 'ga_analyticator_global_notice' );
+
+function ga_analyticator_deactivate() {
+	delete_option('ga_analyticator_global_notification');
+}
+register_deactivation_hook( __FILE__, 'ga_analyticator_deactivate' );
+
+function ga_analyticator_other_plugins_page() {
+	include(plugin_dir_path( __FILE__ ).'/other_plugins.php');
+}
+
+function ga_analyticator_top_level_menu() {
+	$menu_page = add_menu_page( 'Google Analytics', 'Google Analytics', 'manage_options', 'google-analyticator', 'ga_settings_page', 'dashicons-chart-line');
+	add_action('load-'.$menu_page, 'ga_pre_load' );
+	add_submenu_page( 'google-analyticator', 'Other Plugins', 'Other Plugins', 'manage_options', 'google-analyticator-other-plugins', 'ga_analyticator_other_plugins_page');
+
+	$activate_page = add_submenu_page( null, 'Activation', 'Google Analytics', 'manage_options', 'ga_activate' , 'ga_activate');
+	$reset_page = add_submenu_page(null, 'Reset', 'Reset', 'activate_plugins', 'ga_reset', 'ga_reset' );
+   	add_action('load-'.$reset_page, 'ga_do_reset' );
+}
+
+add_action( 'admin_menu', 'ga_analyticator_top_level_menu' );
